+++ nano/node/ipc.hpp
@@ -41,45 +41,6 @@ namespace ipc
 		virtual ~transport () = default;
 	};
 
-	/** Base class for transport configurations */
-	class ipc_config_transport
-	{
-	public:
-		virtual ~ipc_config_transport () = default;
-		bool enabled{ false };
-		size_t io_timeout{ 15 };
-		long io_threads{ -1 };
-	};
-
-	/** Domain socket specific transport config */
-	class ipc_config_domain_socket : public ipc_config_transport
-	{
-	public:
-		/**
-		 * Default domain socket path for Unix systems. Once Boost supports Windows 10 usocks,
-		 * this value will be conditional on OS.
-		 */
-		std::string path{ "/tmp/nano" };
-	};
-
-	/** TCP specific transport config */
-	class ipc_config_tcp_socket : public ipc_config_transport
-	{
-	public:
-		/** Listening port */
-		uint16_t port{ 7077 };
-	};
-
-	/** IPC configuration */
-	class ipc_config
-	{
-	public:
-		nano::error deserialize_json (nano::jsonconfig & json_a);
-		nano::error serialize_json (nano::jsonconfig & json) const;
-		ipc_config_domain_socket transport_domain;
-		ipc_config_tcp_socket transport_tcp;
-	};
-
 	/** The IPC server accepts connections on one or more configured transports */
 	class ipc_server
 	{
